+++ build_runner_core/lib/src/environment/create_merged_dir.dart
@@ -7,11 +7,12 @@ import 'dart:convert';
 import 'dart:io';
 
 import 'package:build/build.dart';
+// ignore: implementation_imports
+import 'package:build/src/internal.dart';
 import 'package:logging/logging.dart';
 import 'package:path/path.dart' as p;
 import 'package:pool/pool.dart';
 
-import '../asset/reader.dart';
 import '../environment/build_environment.dart';
 import '../generate/build_directory.dart';
 import '../generate/finalized_assets_view.dart';
@@ -35,7 +36,7 @@ Future<bool> createMergedOutputDirectories(
     AssetReader reader,
     FinalizedAssetsView finalizedAssetsView,
     bool outputSymlinksOnly) async {
-  if (outputSymlinksOnly && reader is! PathProvidingAssetReader) {
+  if (outputSymlinksOnly && reader.assetPathProvider == null) {
     _logger.severe(
         'The current environment does not support symlinks, but symlinks were '
         'requested.');
@@ -229,7 +230,7 @@ Future<AssetId> _writeAsset(
         await Link(_filePathFor(outputDir, outputId)).create(
             // We assert at the top of `createMergedOutputDirectories` that the
             // reader implements this type when requesting symlinks.
-            (reader as PathProvidingAssetReader).pathTo(id),
+            reader.assetPathProvider!.pathFor(id),
             recursive: true);
       } else {
         await _writeAsBytes(outputDir, outputId, await reader.readAsBytes(id));